Deleted the LogMessageService interface
